ex52 complete

+++ src/ex41/Ex41.java
@@ -2,15 +2,15 @@ package ex41;
 import interfaces.output.Output;
 public class Ex41 {
     public static void main(String[]args){
-int [][]a={{1,2,3,4},{5,6,7,8},{9,10,11,12},{13,14,15,16}};
+double [][]a={{1,2,3,4},{5,6,7,8},{9,10,11,12},{13,14,15,16}};
 Output.printArray(a);
 Output.printArray(intercambiarFilasPorColumnas(a));
 
 }
-public static int [][] intercambiarFilasPorColumnas(int [][] a){
+public static double [][] intercambiarFilasPorColumnas(double [][] a){
         final int aRow=a.length;
         final int aCol=a[0].length;
-        int [][]r= new int[aCol][aRow];
+        double [][]r= new double[aCol][aRow];
       for (int i = 0; i < aRow;i++) {
           for (int e = 0; e < aCol; e++) {
               r[e][i] = a[i][e];
